Add files via upload
If you decide to download this and use it in its raw form there are default passwords I made that are included in the script.  I would recommend changing these in the script for ease of use.  The guide is raw but should get you through the initial set-up.  Only been tested on a closed network.  Would be beneficial if someone has the time to test over two separate networks and include the set-up in the guide.  e.g. port forwarding on the router etc.
This project is just a bit of fun as I learn about secure messaging so please include additional features just update the changes.  

upon initial set-up of clients and servers you are prompted to export .pem files to a USB please have a USB and all required PCs ready.  you WILL HAVE TO CHANGE the exported .pem from intial set-up to e.g. import-server.pem or import-client1.pem all other points were pems have to be downloaded to USBs the name can remain the default and when it is imported on the remote node the .pem will be wiped from the USB.  everything else should be included in the guide.  

This is very raw and designed to only be used in the CLI at this stage would like to eventually have a front-end user friendly web interface at some stage in the future.